refactor: decompose workflow/tools.ts into next-action, cancel, and query modules
Extract handleNextAction + HUMAN_CHECKPOINT_PHASES + PHASE_ACTION_MAP +
findCompoundForPhase into next-action.ts (260L).

Extract handleCancel with saga compensation into cancel.ts (221L).

Extract handleSummary + handleReconcile + handleTransitions into query.ts (255L).

workflow/tools.ts reduced from ~1100 to 412 lines (CRUD + checkpoint only).
